python bindings: re-arranged sub-modules
