+++ python/CMakeLists.txt
@@ -43,13 +43,13 @@ message("  -> site-packages: " ${SP_DIR})
 set(PY_FASTSCAPELIB _fastscapelib_py)
 
 pybind11_add_module(${PY_FASTSCAPELIB}
+    src/algo.cpp
+    src/eroders.cpp
     src/flow_graph.cpp
     src/flow_router.cpp
     src/grid.cpp
     src/main.cpp
-    src/modelings.cpp
-    src/sink_resolver.cpp
-    src/sinks.cpp)
+    src/sink_resolver.cpp)
 
 target_include_directories(${PY_FASTSCAPELIB} PUBLIC ${NUMPY_INCLUDE_DIRS})
