+++ python/fastscapelib/flow/__init__.py
@@ -1,0 +1,1 @@
+from _fastscapelib_py.flow import *
