--- python/fastscapelib/flow_graph/__init__.py
@@ -1,1 +1,0 @@
-from _fastscapelib_py.flow_graph import *
