--- python/fastscapelib/sinks/__init__.py
@@ -1,1 +1,0 @@
-from _fastscapelib_py.sinks import *
